+++ lib/elixir/lib/inspect/algebra.ex
@@ -555,8 +555,7 @@ defmodule Inspect.Algebra do
   @doc since: "1.6.0"
   @spec string(String.t()) :: doc_string
   def string(string) when is_binary(string) do
-    # Use :string for bootstrap reasons
-    doc_string(string, :string.length(string))
+    doc_string(string, String.length(string))
   end
 
   @doc ~S"""
